Add a result? method to Result < Model
self.return? -> h[:result_jsonb].empty? i.e. test whether any result exists in the jsonb data, If not, then the climber hasn't started. Use this when ranking climbers.

Make this a class method not because it operates on the Sequel::Dataset or Sequel::Model but rather on a hash of the result. We may